+++ package.json
@@ -492,7 +492,7 @@
     "test-cypress-host-sample-apps": "tsx ./e2e/runner/run_cypress_host_sample_apps.ts",
     "test-debug": "bun run build:cljs && node --inspect-brk node_modules/.bin/jest --runInBand --detectOpenHandles",
     "test-timezones": "bun install && ./frontend/test/__runner__/run_timezone_tests",
-    "test-memory": "node --expose-gc ./node_modules/.bin/jest --runInBand --selectProjects memory",
+    "test-memory": "node --expose-gc ./node_modules/.bin/jest --runInBand --config jest.memory.conf.js",
     "test-timezones-unit": "jest --silent --maxWorkers=4 --config jest.tz.unit.conf.js",
     "test-unit": "bun run build:cljs && jest --maxWorkers=4",
     "test-unit-keep-cljs": "jest --maxWorkers=4",
